+++ frontend/declarations/main/main.did.d.ts
@@ -25,7 +25,7 @@ export interface DepChange {
   'newVersion' : string,
 }
 export interface DependencyV2 {
-  'name' : PackageName__1,
+  'name' : PackageName,
   'repo' : string,
   'version' : string,
 }
@@ -56,6 +56,8 @@ export interface HttpTransformArg {
   'response' : HttpResponse,
 }
 export interface Main {
+  'addMaintainer' : ActorMethod<[PackageName__1, Principal], Result_3>,
+  'addOwner' : ActorMethod<[PackageName__1, Principal], Result_3>,
   'backup' : ActorMethod<[], undefined>,
   'computeHashesForExistingFiles' : ActorMethod<[], undefined>,
   'finishPublish' : ActorMethod<[PublishingId], Result>,
@@ -63,35 +65,40 @@ export interface Main {
   'getBackupCanisterId' : ActorMethod<[], Principal>,
   'getDefaultPackages' : ActorMethod<
     [string],
-    Array<[PackageName, PackageVersion__1]>
+    Array<[PackageName__1, PackageVersion__1]>
   >,
   'getDownloadTrendByPackageId' : ActorMethod<
     [PackageId],
     Array<DownloadsSnapshot__1>
   >,
   'getDownloadTrendByPackageName' : ActorMethod<
-    [PackageName],
+    [PackageName__1],
     Array<DownloadsSnapshot__1>
   >,
-  'getFileHashes' : ActorMethod<[PackageName, PackageVersion__1], Result_8>,
+  'getFileHashes' : ActorMethod<[PackageName__1, PackageVersion__1], Result_8>,
   'getFileHashesByPackageIds' : ActorMethod<
     [Array<PackageId>],
     Array<[PackageId, Array<[FileId, Uint8Array | number[]]>]>
   >,
   'getFileHashesQuery' : ActorMethod<
-    [PackageName, PackageVersion__1],
+    [PackageName__1, PackageVersion__1],
     Result_8
   >,
-  'getFileIds' : ActorMethod<[PackageName, PackageVersion__1], Result_7>,
+  'getFileIds' : ActorMethod<[PackageName__1, PackageVersion__1], Result_7>,
   'getHighestSemverBatch' : ActorMethod<
-    [Array<[PackageName, PackageVersion__1, SemverPart]>],
+    [Array<[PackageName__1, PackageVersion__1, SemverPart]>],
     Result_6
   >,
-  'getHighestVersion' : ActorMethod<[PackageName], Result_5>,
+  'getHighestVersion' : ActorMethod<[PackageName__1], Result_5>,
   'getMostDownloadedPackages' : ActorMethod<[], Array<PackageSummary>>,
   'getMostDownloadedPackagesIn7Days' : ActorMethod<[], Array<PackageSummary>>,
   'getNewPackages' : ActorMethod<[], Array<PackageSummary>>,
-  'getPackageDetails' : ActorMethod<[PackageName, PackageVersion__1], Result_4>,
+  'getPackageDetails' : ActorMethod<
+    [PackageName__1, PackageVersion__1],
+    Result_4
+  >,
+  'getPackageMaintainers' : ActorMethod<[PackageName__1], Array<Principal>>,
+  'getPackageOwners' : ActorMethod<[PackageName__1], Array<Principal>>,
   'getPackagesByCategory' : ActorMethod<
     [],
     Array<[string, Array<PackageSummary>]>
@@ -105,23 +112,24 @@ export interface Main {
   'getTotalPackages' : ActorMethod<[], bigint>,
   'getUser' : ActorMethod<[Principal], [] | [User__1]>,
   'http_request' : ActorMethod<[Request], Response>,
-  'notifyInstall' : ActorMethod<[PackageName, PackageVersion__1], undefined>,
+  'notifyInstall' : ActorMethod<[PackageName__1, PackageVersion__1], undefined>,
   'notifyInstalls' : ActorMethod<
-    [Array<[PackageName, PackageVersion__1]>],
+    [Array<[PackageName__1, PackageVersion__1]>],
     undefined
   >,
+  'removeMaintainer' : ActorMethod<[PackageName__1, Principal], Result_3>,
+  'removeOwner' : ActorMethod<[PackageName__1, Principal], Result_3>,
   'restore' : ActorMethod<[bigint], undefined>,
   'search' : ActorMethod<
     [Text, [] | [bigint], [] | [bigint]],
     [Array<PackageSummary>, PageCount]
   >,
-  'setUserProp' : ActorMethod<[string, string], Result_1>,
+  'setUserProp' : ActorMethod<[string, string], Result_3>,
   'startFileUpload' : ActorMethod<
     [PublishingId, Text, bigint, Uint8Array | number[]],
-    Result_3
+    Result_2
   >,
-  'startPublish' : ActorMethod<[PackageConfigV3_Publishing], Result_2>,
-  'transferOwnership' : ActorMethod<[PackageName, Principal], Result_1>,
+  'startPublish' : ActorMethod<[PackageConfigV3_Publishing], Result_1>,
   'transformRequest' : ActorMethod<[HttpTransformArg], HttpResponse>,
   'uploadBenchmarks' : ActorMethod<[PublishingId, Benchmarks], Result>,
   'uploadFileChunk' : ActorMethod<
@@ -145,7 +153,7 @@ export interface PackageConfigV3 {
   'scripts' : Array<Script>,
   'baseDir' : string,
   'documentation' : string,
-  'name' : PackageName__1,
+  'name' : PackageName,
   'homepage' : string,
   'description' : string,
   'version' : string,
@@ -164,7 +172,7 @@ export interface PackageConfigV3_Publishing {
   'scripts' : Array<Script>,
   'baseDir' : string,
   'documentation' : string,
-  'name' : PackageName__1,
+  'name' : PackageName,
   'homepage' : string,
   'description' : string,
   'version' : string,
@@ -180,10 +188,13 @@ export interface PackageConfigV3_Publishing {
 export interface PackageDetails {
   'benchmarks' : Benchmarks__1,
   'ownerInfo' : User,
+  'owners' : Array<User>,
+  'maintainers' : Array<User>,
   'owner' : Principal,
   'depAlias' : string,
   'deps' : Array<PackageSummary__1>,
   'quality' : PackageQuality,
+  'publisher' : User,
   'testStats' : TestStats__1,
   'highestVersion' : PackageVersion,
   'downloadsTotal' : bigint,
@@ -222,9 +233,12 @@ export interface PackageQuality {
 }
 export interface PackageSummary {
   'ownerInfo' : User,
+  'owners' : Array<User>,
+  'maintainers' : Array<User>,
   'owner' : Principal,
   'depAlias' : string,
   'quality' : PackageQuality,
+  'publisher' : User,
   'highestVersion' : PackageVersion,
   'downloadsTotal' : bigint,
   'downloadsInLast30Days' : bigint,
@@ -234,9 +248,12 @@ export interface PackageSummary {
 }
 export interface PackageSummaryWithChanges {
   'ownerInfo' : User,
+  'owners' : Array<User>,
+  'maintainers' : Array<User>,
   'owner' : Principal,
   'depAlias' : string,
   'quality' : PackageQuality,
+  'publisher' : User,
   'highestVersion' : PackageVersion,
   'downloadsTotal' : bigint,
   'downloadsInLast30Days' : bigint,
@@ -247,9 +264,12 @@ export interface PackageSummaryWithChanges {
 }
 export interface PackageSummaryWithChanges__1 {
   'ownerInfo' : User,
+  'owners' : Array<User>,
+  'maintainers' : Array<User>,
   'owner' : Principal,
   'depAlias' : string,
   'quality' : PackageQuality,
+  'publisher' : User,
   'highestVersion' : PackageVersion,
   'downloadsTotal' : bigint,
   'downloadsInLast30Days' : bigint,
@@ -260,9 +280,12 @@ export interface PackageSummaryWithChanges__1 {
 }
 export interface PackageSummary__1 {
   'ownerInfo' : User,
+  'owners' : Array<User>,
+  'maintainers' : Array<User>,
   'owner' : Principal,
   'depAlias' : string,
   'quality' : PackageQuality,
+  'publisher' : User,
   'highestVersion' : PackageVersion,
   'downloadsTotal' : bigint,
   'downloadsInLast30Days' : bigint,
@@ -291,17 +314,17 @@ export interface Response {
 }
 export type Result = { 'ok' : null } |
   { 'err' : Err };
-export type Result_1 = { 'ok' : null } |
+export type Result_1 = { 'ok' : PublishingId } |
+  { 'err' : Err };
+export type Result_2 = { 'ok' : FileId } |
+  { 'err' : Err };
+export type Result_3 = { 'ok' : null } |
   { 'err' : string };
-export type Result_2 = { 'ok' : PublishingId } |
-  { 'err' : Err };
-export type Result_3 = { 'ok' : FileId } |
-  { 'err' : Err };
 export type Result_4 = { 'ok' : PackageDetails } |
   { 'err' : Err };
 export type Result_5 = { 'ok' : PackageVersion__1 } |
   { 'err' : Err };
-export type Result_6 = { 'ok' : Array<[PackageName, PackageVersion__1]> } |
+export type Result_6 = { 'ok' : Array<[PackageName__1, PackageVersion__1]> } |
   { 'err' : Err };
 export type Result_7 = { 'ok' : Array<FileId> } |
   { 'err' : Err };